==> notebook/08 Calculate User Rating.ipynb <==
# FIle 08: Calculating User Rating

### Input Files:
- db/07-timeline-tweets-with-thresholding.csv

### Output Files:
- db/08-user-rating.csv

### Steps:
1. import required libraries
1. loading preprocessed dataset
1. make of list all the index for each tweet for each user
1. create function to rate the user
1. caluclate user rating
1. create dataframe
1. save dataframe

# import required libraries
import numpy as np
import pandas as pd
from tqdm import tqdm

# loading preprocessed dataset...
df = pd.read_csv("../db/07-timeline-tweets-with-thresholding.csv")
username = list(set(df.USER.tolist()))

# make of list all the index for each tweet for each user
rows = []
for name in tqdm(username) :
    rows.append(df.loc[df['USER'] == name].index)

# create function to rate the user
def calculate_rating(index) :
    index = rows[index]
    thresh = []
    for x in index :
        thresh.append(df['THRESHOLD'][x])

    calc = (thresh.count(3) - thresh.count(0))
    return 0 if calc < 0 else 1

# caluclate user rating
rating = []
for x in tqdm(range(len(username))) :
    rating.append(calculate_rating(x))

# create dataframe
final = pd.DataFrame(
    list(zip(username, rating)),
    columns =['USER', 'RATING']
)

# save dataframe
final.to_csv("../db/08-user-rating.csv", index=False)
